+++ src/database/database.go
@@ -1,0 +1,21 @@
+package controllers
+
+import (
+	"fmt"
+
+	model "github.com/ITU-DevOps-N/go-minitwit/src/models"
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
+)
+
+var DB *gorm.DB
+
+func SetupDB(database string) {
+	db, err := gorm.Open(sqlite.Open(database), &gorm.Config{})
+	if err != nil {
+		fmt.Printf("Error: %s", err.Error())
+		panic("Failed to connect to database.")
+	}
+	db.AutoMigrate(&model.User{}, &model.Message{}, &model.Follow{})
+	DB = db
+}
