+++ src/web/templates/register.tpl
